+++ FyreWorksPM/Services/BidLineItem/TrimTimeLookupService.cs
@@ -2,16 +2,16 @@
 {
     public static class TrimTimeLookupService
     {
-        private static readonly Dictionary<string, double> _trimHoursByType =
+        private static readonly Dictionary<string, decimal> _trimHoursByType =
             new()
             {
-                ["normal"] = 0.5,
-                ["lift"] = 1.0,
-                ["panel"] = 1.0,
-                ["pipe"] = 1.0
+                ["normal"] = 0.5m,
+                ["lift"] = 1.0m,
+                ["panel"] = 1.0m,
+                ["pipe"] = 1.0m
             };
 
-        public static double GetHours(string installType)
+        public static decimal GetHours(string installType)
         {
             if (string.IsNullOrWhiteSpace(installType))
                 return 0;
